pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lore-sdk"
version = "0.12.0"
description = "Cross-agent semantic memory with knowledge graphs, fact extraction, and MCP integration"
readme = "README.md"
license = "MIT"
authors = [{ name = "Amit Paz", email = "amit.paz@gmail.com" }]
keywords = [
    "agent", "memory", "lore", "knowledge",
    "knowledge-graph", "fact-extraction", "memory-consolidation",
    "mcp", "model-context-protocol",
    "cognitive-memory", "ai-memory", "semantic-memory", "agent-memory",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "python-ulid>=2.0.0",
    "numpy>=1.21.0",
    "onnxruntime>=1.14.0",
    "tokenizers>=0.13.0",
]

[project.optional-dependencies]
remote = [
    "httpx>=0.24.0",
]
mcp = [
    "mcp>=1.0.0",
]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "asyncpg>=0.28.0",
    "pydantic>=2.0.0",
    "PyJWT[crypto]>=2.8",
]
redis = [
    "redis>=4.5.0",
]
aws = [
    "boto3>=1.26.0",
]
security = [
    "detect-secrets>=1.4.0",
]
ner = [
    "spacy>=3.5.0",
]
enrichment = [
    "litellm>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
]

[project.scripts]
lore = "lore.cli:main"

[project.urls]
Homepage = "https://github.com/amitpaz1/lore"
Repository = "https://github.com/amitpaz1/lore"
Documentation = "https://github.com/amitpaz1/lore/tree/main/docs"
Issues = "https://github.com/amitpaz1/lore/issues"
Changelog = "https://github.com/amitpaz1/lore/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/lore"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires Docker Compose stack (deselect with -m 'not integration')",
]

[tool.ruff]
target-version = "py39"
src = ["src"]
line-length = 180

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E741", "N815", "N812"]
